Add Play Controls launch guard settings
